requirements.txt
opencv-contrib-python==4.10.0.84
numpy==1.26.4
pillow==11.0.0